+++ controllers/category.php
@@ -25,10 +25,10 @@ function category()
     }
 
 
-    $oid = $mysqli->query("SELECT p.product_id, p.product_name, p.price, pv.variant_id, pi.file_name
-                                                FROM product p JOIN product_variant pv ON (p.product_id = pv.product_id) JOIN product_image pi ON (pv.variant_id = pi.variant_id)
+    $oid = $mysqli->query("SELECT p.product_id, p.product_name, p.price, pi.file_name
+                                                FROM product p JOIN product_image pi ON (pi.product_id = p.product_id)
                                                 WHERE p.quantity_available > 0 AND p.category_id = $id
-                                                  AND (pv.default = true AND pi.type = 'main')
+                                                AND pi.type = 'main'
                                                 ORDER BY p.product_name");
 
     if ($oid->num_rows == 0) {